refactor(store): split app store and transforms
