+++ src/lib/drawing/transform/index.ts
@@ -1,0 +1,6 @@
+export * from './resize';
+export * from './rotate';
+export * from './crop';
+export * from './move';
+export * from './flip';
+export * from './scale';
